requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.7.2
black==24.4.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==24.3
django-sslserver==0.22
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
openai==1.35.3
opencv-python-headless==4.9.0.80
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
pycparser==2.21
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.1
wheel==0.43.0
